rename stacker to flex-stacker
