+++ .github/workflows/stacker.yaml
@@ -2,10 +2,10 @@ name: 'Stacker build/test'
 on:
   pull_request:
     paths:
-      - 'stm32-modules/stacker/**/*'
+      - 'stm32-modules/flex-stacker/**/*'
       - 'stm32-modules/common/**/*'
       - 'stm32-modules/include/common/**/*'
-      - 'stm32-modules/include/stacker/**/*'
+      - 'stm32-modules/include/flex-stacker/**/*'
       - 'cmake/**/*'
       - 'CMakeLists.txt'
       - 'stm32-modules/CMakeLists.txt'
@@ -17,23 +17,23 @@ on:
       - 'cmake/Arduino*'
   push:
     paths:
-      - 'stm32-modules/stacker/**/*'
+      - 'stm32-modules/flex-stacker/**/*'
       - 'stm32-modules/common/**/*'
       - 'stm32-modules/include/common/**/*'
-      - 'stm32-modules/include/stacker/**/*'
+      - 'stm32-modules/include/flex-stacker/**/*'
       - 'cmake/**/*'
       - 'CMakeLists.txt'
       - 'stm32-modules/CMakeLists.txt'
       - 'CMakePresets.json'
       - '.clang-format'
       - '.clang-tidy'
-      - '.github/workflows/stacker.yaml'
+      - '.github/workflows/flex-stacker.yaml'
     paths_ignore:
       - 'cmake/Arduino*'
     branches:
       - '*'
     tags:
-      - 'stacker@*'
+      - 'flex-stacker@*'
   workflow_dispatch:
 
 
@@ -59,17 +59,17 @@ jobs:
       - name: 'Configure'
         run: cmake --preset=stm32-cross .
       - name: 'Format'
-        run: cmake --build --preset cross --target stacker-format-ci
+        run: cmake --build --preset cross --target flex-stacker-format-ci
       - name: 'Lint'
-        run: cmake --build --preset cross --target stacker-lint
+        run: cmake --build --preset cross --target flex-stacker-lint
       - name: 'Build startup app'
-        run: cmake --build --preset cross --target stacker-startup
+        run: cmake --build --preset cross --target flex-stacker-startup
       - name: 'Build'
-        run: cmake --build --preset cross --target stacker-hex
+        run: cmake --build --preset cross --target flex-stacker-hex
       - name: 'Build full image'
-        run: cmake --build --preset cross --target stacker-image-hex
+        run: cmake --build --preset cross --target flex-stacker-image-hex
       - name: 'Build full image binary'
-        run: cmake --build --preset cross --target stacker-image-bin
+        run: cmake --build --preset cross --target flex-stacker-image-bin
   host-compile-test:
     name: 'Host-Compile/Test'
     runs-on: 'ubuntu-20.04'
@@ -92,6 +92,6 @@ jobs:
       - name: 'Configure'
         run: cmake --preset=stm32-host .
       - name: 'Build Simulator'
-        run: cmake --build --preset host --target stacker-simulator
+        run: cmake --build --preset host --target flex-stacker-simulator
       - name: 'Build and Test'
-        run: cmake --build --preset stacker-tests
+        run: cmake --build --preset flex-stacker-tests
