+++ stm32-modules/CMakeLists.txt
@@ -36,6 +36,6 @@ add_subdirectory(common)
 add_subdirectory(heater-shaker)
 add_subdirectory(thermocycler-gen2)
 add_subdirectory(tempdeck-gen3)
-add_subdirectory(stacker)
+add_subdirectory(flex-stacker)
 
 coverage_evaluate()
