+++ stm32-modules/flex-stacker/firmware/host_comms_task/usb_hardware.c
