+++ stm32-modules/flex-stacker/firmware/host_comms_task/usbd_desc.h
