+++ stm32-modules/flex-stacker/firmware/motor_control/motor_hardware.c
