+++ stm32-modules/flex-stacker/firmware/system/stm32g4xx_hal_msp.c
