+++ stm32-modules/flex-stacker/firmware/system/system_stm32g4xx.c
