+++ stm32-modules/flex-stacker/src/errors.cpp
@@ -1,4 +1,4 @@
-#include "stacker/errors.hpp"
+#include "flex-stacker/errors.hpp"
 
 using namespace errors;
