+++ stm32-modules/flex-stacker/src/version.cpp.in
