+++ stm32-modules/include/flex-stacker/firmware/usb_hardware.h
