+++ stm32-modules/include/flex-stacker/flex-stacker/errors.hpp
